convert smiTest1 to catch
